+++ mne/channels/montage.py
@@ -25,7 +25,8 @@ from ..io.meas_info import (_make_dig_points, _read_dig_points, _read_dig_fif,
 from ..io.pick import pick_types
 from ..io.open import fiff_open
 from ..io.constants import FIFF
-from ..utils import _check_fname, warn, copy_function_doc_to_method_doc
+from ..utils import (_check_fname, warn, copy_function_doc_to_method_doc,
+                     _clean_names)
 
 from ..externals.six import string_types
 from ..externals.six.moves import map
@@ -184,6 +185,11 @@ def read_montage(kind, ch_names=None, path=None, unit='m', transform=False):
     GSN-HydroCel-257      HydroCel Geodesic Sensor Net and Cz (257+3 locations)
 
     10-5_EGI129           462+3 locations
+
+    mgh60                 The (older) 60-channel cap used at
+                          MGH (60+3 locations)
+    mgh70                 The (newer) 70-channel BrainVision cap used at
+                          MGH (70+3 locations)
     ===================   =====================================================
 
     .. versionadded:: 0.9.0
@@ -794,6 +800,9 @@ def _set_montage(info, montage, update_ch_names=False):
         else:
             montage_ch_names = montage.ch_names
             info_ch_names = info['ch_names']
+        info_ch_names = _clean_names(info_ch_names, remove_whitespace=True)
+        montage_ch_names = _clean_names(montage_ch_names,
+                                        remove_whitespace=True)
 
         for pos, ch_name in zip(montage.pos, montage_ch_names):
             if ch_name not in info_ch_names: